refactor : 위치 변경
