+++ src/main/java/domain/lottos/LottoStatistics.java
@@ -1,4 +1,4 @@
-package domain.lottoMatcher;
+package domain.lottos;
 
 import java.util.Map;
